fix(guide-booking): cast averageRating to ::float
postgres.js returns PostgreSQL numeric as a string. The sql<number>
annotation is TypeScript-only and has no runtime effect. Adding ::float
ensures the driver returns a JS number, keeping the API contract valid.

+++ services/guide-booking/src/stats/stats.service.ts
@@ -51,7 +51,7 @@ export class StatsService {
       this.db
         .select({
           total: count(),
-          averageRating: sql<number>`COALESCE(ROUND(AVG(${guideReviews.rating})::numeric, 1), 0)`,
+          averageRating: sql<number>`COALESCE(ROUND(AVG(${guideReviews.rating})::numeric, 1), 0)::float`,
         })
         .from(guideReviews)
         .where(eq(guideReviews.isActive, true)),
